week-2/requirements.txt
functions-framework==3.*
google-cloud-storage 